LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-multiple-origins-worker-expected.txt

CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.

FAIL 3 origins allowed, match the 3rd (http://localhost:8800) promise_test: Unhandled rejection with value: object "TypeError: Access-Control-Allow-Origin cannot contain more than one origin."
FAIL 3 origins allowed, match the 3rd ("*") promise_test: Unhandled rejection with value: object "TypeError: Access-Control-Allow-Origin cannot contain more than one origin."
FAIL 3 origins allowed, match twice (http://localhost:8800) promise_test: Unhandled rejection with value: object "TypeError: Access-Control-Allow-Origin cannot contain more than one origin."
FAIL 3 origins allowed, match twice ("*") promise_test: Unhandled rejection with value: object "TypeError: Access-Control-Allow-Origin cannot contain more than one origin."
FAIL 3 origins allowed, match twice ("*" and http://localhost:8800) promise_test: Unhandled rejection with value: object "TypeError: Access-Control-Allow-Origin cannot contain more than one origin."
PASS 3 origins allowed, no match 
